Add state to a component - 2
Here we add a state to a Functional component.
Remember Funcional component are stateless. So we functional component are called, the parent or ancestor to the component has to pass a state to the Functional component it is calling.

FYI: The previous "Add a state commit" was more of initializing a state for a class component. Class component are not stateless. The must have a state at initialization(this means when the class must have a constuctor that initializes the state.)