Update State - 2
Here we have a component(DisplayQuizz) that calls a handleAnswer arrow method in the parent or ancestor component though the helps of props it was able to reference the function. This arrow function updates the state of the parent function.

For the App coponent, it defines its own internal state and also alters its own state but it alters its state with the help of the of of its child component(DisplayQuizz)